requirements.txt
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
# cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
flasgger==0.9.5
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-paginate==2022.1.8
Flask-WTF==1.0.0
# greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.15.0
mistune==2.0.2
mongoengine==0.23.1
packaging==21.3
passlib==1.7.4
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.7
pyrsistent==0.18.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
